+++ src/main/java/com/example/antiepidemic/service/ImageService.java
@@ -15,7 +15,7 @@ public interface ImageService {
      * @param inputer 管理员id
      * @return java.lang.Boolean
      */
-    public Boolean insertImage(String name,String path,Integer inputer);
+    public Boolean insertImage(String name,String path,Long inputer);
 
     /**
      * 修改图片
@@ -24,21 +24,21 @@ public interface ImageService {
      * @param path 图片路径
      * @return java.lang.Boolean
      */
-    public Boolean updateImage(Integer id,String name,String path);
+    public Boolean updateImage(Long id,String name,String path);
 
     /**
      * 删除图片
      * @param id 图片id
      * @return java.lang.Boolean
      */
-    public Boolean deleteImage(Integer id);
+    public Boolean deleteImage(Long id);
 
     /**
      * 获取单个图片
      * @param id 图片id
      * @return com.example.antiepidemic.domain.Image
      */
-    public Image queryOneImage(Integer id);
+    public Image queryOneImage(Long id);
 
     /**
      * 获取全部图片
